+++ .github/workflows/test.yml
@@ -732,7 +732,9 @@ jobs:
             './gef/cypress/e2e/eligibility/**/*.spec.js',
             './gef/cypress/e2e/exporter/**/*.spec.js',
             './gef/cypress/e2e/external/**/*.spec.js',
-            './gef/cypress/e2e/facility/**/*.spec.js',
+            './gef/cypress/e2e/facility/*.spec.js',
+            './gef/cypress/e2e/facility/unissued-facilities/AIN/*.spec.js',
+            './gef/cypress/e2e/facility/unissued-facilities/MIN/*.spec.js',
             './gef/cypress/e2e/maker/**/*.spec.js',
             './gef/cypress/e2e/scenarios/**/*.spec.js',
             './gef/cypress/e2e/security/**/*.spec.js',
