+++ e2e-tests/gef/cypress/e2e/facility/unissued-facilities/AIN/unissued-facilities-facility-end-date-AIN.spec.js
@@ -1,19 +1,19 @@
-import relative from '../../relativeURL';
-import { errorSummary } from '../../partials';
-import CONSTANTS from '../../../fixtures/constants';
-import dateConstants from '../../../../../e2e-fixtures/dateConstants';
-import { MOCK_APPLICATION_AIN } from '../../../fixtures/mocks/mock-deals';
-import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
-import { MOCK_FACILITY_ONE, MOCK_FACILITY_TWO, MOCK_FACILITY_THREE, MOCK_FACILITY_FOUR } from '../../../fixtures/mocks/mock-facilities';
-import unissuedFacilityTable from '../../pages/unissued-facilities';
-import applicationPreview from '../../pages/application-preview';
-import aboutFacilityUnissued from '../../pages/unissued-facilities-about-facility';
-import facilityEndDate from '../../pages/facility-end-date';
+import relative from '../../../relativeURL';
+import { errorSummary } from '../../../partials';
+import CONSTANTS from '../../../../fixtures/constants';
+import dateConstants from '../../../../../../e2e-fixtures/dateConstants';
+import { MOCK_APPLICATION_AIN } from '../../../../fixtures/mocks/mock-deals';
+import { BANK1_MAKER1 } from '../../../../../../e2e-fixtures/portal-users.fixture';
+import { MOCK_FACILITY_ONE, MOCK_FACILITY_TWO, MOCK_FACILITY_THREE, MOCK_FACILITY_FOUR } from '../../../../fixtures/mocks/mock-facilities';
+import unissuedFacilityTable from '../../../pages/unissued-facilities';
+import applicationPreview from '../../../pages/application-preview';
+import aboutFacilityUnissued from '../../../pages/unissued-facilities-about-facility';
+import facilityEndDate from '../../../pages/facility-end-date';
 
 const facilityEndDateEnabled = Number(Cypress.env('GEF_DEAL_VERSION')) >= 1;
 
 if (facilityEndDateEnabled) {
-  context('Unissued Facilities - facility end date page', () => {
+  context('Unissued Facilities AIN - facility end date page', () => {
     let dealId;
     let token;
     let facilityOneId;
