+++ e2e-tests/gef/cypress/e2e/facility/unissued-facilities/MIN/unissued-facility-change-from-preview-specialIssuedPermission-MIN.spec.js
@@ -1,28 +1,38 @@
 import { format } from 'date-fns';
 
-import relative from '../../relativeURL';
+import relative from '../../../relativeURL';
 
-import CONSTANTS from '../../../fixtures/constants';
+import CONSTANTS from '../../../../fixtures/constants';
 
-import dateConstants from '../../../../../e2e-fixtures/dateConstants';
+import dateConstants from '../../../../../../e2e-fixtures/dateConstants';
 
-import { MOCK_APPLICATION_MIN } from '../../../fixtures/mocks/mock-deals';
-import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
-import { MOCK_FACILITY_ONE, MOCK_FACILITY_TWO, MOCK_FACILITY_THREE, MOCK_FACILITY_FOUR } from '../../../fixtures/mocks/mock-facilities';
+import { MOCK_APPLICATION_MIN } from '../../../../fixtures/mocks/mock-deals';
+import { BANK1_MAKER1 } from '../../../../../../e2e-fixtures/portal-users.fixture';
+import { MOCK_FACILITY_ONE, MOCK_FACILITY_TWO, MOCK_FACILITY_THREE, MOCK_FACILITY_FOUR } from '../../../../fixtures/mocks/mock-facilities';
 
-import { mainHeading, continueButton, errorSummary } from '../../partials';
-import applicationPreview from '../../pages/application-preview';
-import unissuedFacilityTable from '../../pages/unissued-facilities';
-import aboutFacilityUnissued from '../../pages/unissued-facilities-about-facility';
-import applicationSubmission from '../../pages/application-submission';
-import statusBanner from '../../pages/application-status-banner';
-import facilityEndDate from '../../pages/facility-end-date';
+import { continueButton, mainHeading } from '../../../partials';
+import applicationPreview from '../../../pages/application-preview';
+import unissuedFacilityTable from '../../../pages/unissued-facilities';
+import aboutFacilityUnissued from '../../../pages/unissued-facilities-about-facility';
+import applicationSubmission from '../../../pages/application-submission';
+import statusBanner from '../../../pages/application-status-banner';
+import facilityEndDate from '../../../pages/facility-end-date';
 
 let dealId;
 let token;
 let facilityOneId;
 
-const unissuedFacilitiesArray = [MOCK_FACILITY_ONE, MOCK_FACILITY_THREE, MOCK_FACILITY_FOUR];
+const FACILITY_ONE_SPECIAL = { ...MOCK_FACILITY_ONE };
+const FACILITY_TWO_SPECIAL = { ...MOCK_FACILITY_TWO };
+const FACILITY_THREE_SPECIAL = { ...MOCK_FACILITY_THREE };
+const FACILITY_FOUR_SPECIAL = { ...MOCK_FACILITY_FOUR };
+
+FACILITY_ONE_SPECIAL.specialIssuePermission = true;
+FACILITY_TWO_SPECIAL.specialIssuePermission = true;
+FACILITY_THREE_SPECIAL.specialIssuePermission = true;
+FACILITY_FOUR_SPECIAL.specialIssuePermission = true;
+
+const unissuedFacilitiesArray = [FACILITY_ONE_SPECIAL, FACILITY_THREE_SPECIAL, FACILITY_FOUR_SPECIAL];
 
 const facilityEndDateEnabled = Number(Cypress.env('GEF_DEAL_VERSION')) >= 1;
 
@@ -30,7 +40,7 @@ const facilityEndDateEnabled = Number(Cypress.env('GEF_DEAL_VERSION')) >= 1;
   for changing facilities to issued from preview page.
   To unlock functionality, need to first issue one facility from unissued-facility table
 */
-context('Unissued Facilities MIN - change to issued from preview page', () => {
+context('Unissued Facilities MIN - change to issued from preview page - specialIssuedPermission', () => {
   before(() => {
     cy.apiLogin(BANK1_MAKER1)
       .then((t) => {
@@ -42,16 +52,16 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
           cy.apiUpdateApplication(dealId, token, MOCK_APPLICATION_MIN).then(() => {
             cy.apiCreateFacility(dealId, CONSTANTS.FACILITY_TYPE.CASH, token).then((facility) => {
               facilityOneId = facility.body.details._id;
-              cy.apiUpdateFacility(facility.body.details._id, token, MOCK_FACILITY_ONE);
+              cy.apiUpdateFacility(facility.body.details._id, token, FACILITY_ONE_SPECIAL);
             });
             cy.apiCreateFacility(dealId, CONSTANTS.FACILITY_TYPE.CASH, token).then((facility) =>
-              cy.apiUpdateFacility(facility.body.details._id, token, MOCK_FACILITY_TWO),
+              cy.apiUpdateFacility(facility.body.details._id, token, FACILITY_TWO_SPECIAL),
             );
             cy.apiCreateFacility(dealId, CONSTANTS.FACILITY_TYPE.CONTINGENT, token).then((facility) =>
-              cy.apiUpdateFacility(facility.body.details._id, token, MOCK_FACILITY_THREE),
+              cy.apiUpdateFacility(facility.body.details._id, token, FACILITY_THREE_SPECIAL),
             );
             cy.apiCreateFacility(dealId, CONSTANTS.FACILITY_TYPE.CASH, token).then((facility) =>
-              cy.apiUpdateFacility(facility.body.details._id, token, MOCK_FACILITY_FOUR),
+              cy.apiUpdateFacility(facility.body.details._id, token, FACILITY_FOUR_SPECIAL),
             );
             cy.apiSetApplicationStatus(dealId, token, CONSTANTS.DEAL_STATUS.UKEF_ACKNOWLEDGED);
           });
@@ -104,7 +114,7 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
       }
     });
 
-    it('should be able to update facility and then go back to application preview page', () => {
+    it('should be able to update facility and then go back to application preview page with coverStartDate more than 3 months in the future', () => {
       applicationPreview.unissuedFacilitiesReviewLink().click();
       unissuedFacilityTable.updateIndividualFacilityButton(0).click();
 
@@ -124,30 +134,12 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
       if (facilityEndDateEnabled) {
         aboutFacilityUnissued.isUsingFacilityEndDateYes().click();
       }
-
-      cy.clickContinueButton();
-
-      errorSummary().contains('The cover start date must be within 3 months of the inclusion notice submission date');
-      aboutFacilityUnissued.coverStartDateError().contains('The cover start date must be within 3 months of the inclusion notice submission date');
-
-      aboutFacilityUnissued.issueDateDay().clear().type(dateConstants.todayDay);
-      aboutFacilityUnissued.issueDateMonth().clear().type(dateConstants.todayMonth);
-      aboutFacilityUnissued.issueDateYear().clear().type(dateConstants.todayYear);
-
-      aboutFacilityUnissued.shouldCoverStartOnSubmissionNo().click();
-      aboutFacilityUnissued.coverStartDateDay().clear().type(dateConstants.twoMonthsDay);
-      aboutFacilityUnissued.coverStartDateMonth().clear().type(dateConstants.twoMonthsMonth);
-      aboutFacilityUnissued.coverStartDateYear().clear().type(dateConstants.twoMonthsYear);
-
-      aboutFacilityUnissued.coverEndDateDay().clear().type(dateConstants.threeMonthsOneDayDay);
-      aboutFacilityUnissued.coverEndDateMonth().clear().type(dateConstants.threeMonthsOneDayMonth);
-      aboutFacilityUnissued.coverEndDateYear().clear().type(dateConstants.threeMonthsOneDayYear);
       cy.clickContinueButton();
 
       if (facilityEndDateEnabled) {
-        facilityEndDate.facilityEndDateDay().clear().type(dateConstants.threeMonthsOneDayDay);
-        facilityEndDate.facilityEndDateMonth().clear().type(dateConstants.threeMonthsOneDayMonth);
-        facilityEndDate.facilityEndDateYear().clear().type(dateConstants.threeMonthsOneDayYear);
+        facilityEndDate.facilityEndDateDay().clear().type(dateConstants.threeYearsDay);
+        facilityEndDate.facilityEndDateMonth().clear().type(dateConstants.threeYearsMonth);
+        facilityEndDate.facilityEndDateYear().clear().type(dateConstants.threeYearsYear);
         cy.clickContinueButton();
       }
 
@@ -171,8 +163,8 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
 
     it('facility table should have change links on the changed to issued facilities', () => {
       const issuedDate = format(dateConstants.today, 'd MMMM yyyy');
-      const coverStart = format(dateConstants.twoMonths, 'd MMMM yyyy');
-      const coverEnd = format(dateConstants.threeMonthsOneDay, 'd MMMM yyyy');
+      const coverStart = format(dateConstants.twoYears, 'd MMMM yyyy');
+      const coverEnd = format(dateConstants.threeYears, 'd MMMM yyyy');
 
       // can change facility one name and issue dates etc since changed to issued
       applicationPreview.facilitySummaryListTable(3).nameValue().contains(MOCK_FACILITY_ONE.name);
@@ -208,10 +200,6 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
       applicationPreview.facilitySummaryListTable(0).coverStartDateAction().should('not.exist');
       applicationPreview.facilitySummaryListTable(0).coverEndDateAction().should('not.exist');
 
-      if (facilityEndDateEnabled) {
-        applicationPreview.facilitySummaryListTable(0).isUsingFacilityEndDateAction().should('have.class', 'govuk-!-display-none');
-      }
-
       // should not be able to change facility two has previously issued
       applicationPreview.facilitySummaryListTable(2).nameValue().contains(MOCK_FACILITY_TWO.name);
       applicationPreview.facilitySummaryListTable(2).nameAction().should('have.class', 'govuk-!-display-none');
@@ -223,9 +211,13 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
       applicationPreview.facilitySummaryListTable(2).issueDateAction().should('not.exist');
       applicationPreview.facilitySummaryListTable(2).coverStartDateValue().contains('Date you submit the notice');
       applicationPreview.facilitySummaryListTable(2).coverStartDateAction().should('have.class', 'govuk-!-display-none');
+
+      if (facilityEndDateEnabled) {
+        applicationPreview.facilitySummaryListTable(2).isUsingFacilityEndDateAction().should('have.class', 'govuk-!-display-none');
+      }
     });
 
-    it('change unissued to issued from application preview page', () => {
+    it('change unissued to issued from application preview page with coverStartDate more than 3 months in the future', () => {
       // to change to issued from preview page by clicking change on issued row
       applicationPreview.facilitySummaryListTable(0).hasBeenIssuedAction().click();
       aboutFacilityUnissued.facilityName().clear();
@@ -235,31 +227,11 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
       aboutFacilityUnissued.issueDateMonth().type(dateConstants.todayMonth);
       aboutFacilityUnissued.issueDateYear().type(dateConstants.todayYear);
 
-      aboutFacilityUnissued.shouldCoverStartOnSubmissionNo().click();
-      aboutFacilityUnissued.coverStartDateDay().type(dateConstants.twoYearsDay);
-      aboutFacilityUnissued.coverStartDateMonth().type(dateConstants.twoYearsMonth);
-      aboutFacilityUnissued.coverStartDateYear().type(dateConstants.twoYearsYear);
+      aboutFacilityUnissued.shouldCoverStartOnSubmissionYes().click();
 
       aboutFacilityUnissued.coverEndDateDay().type(dateConstants.threeYearsDay);
       aboutFacilityUnissued.coverEndDateMonth().type(dateConstants.threeYearsMonth);
       aboutFacilityUnissued.coverEndDateYear().type(dateConstants.threeYearsYear);
-      cy.clickContinueButton();
-
-      errorSummary().contains('The cover start date must be within 3 months of the inclusion notice submission date');
-      aboutFacilityUnissued.coverStartDateError().contains('The cover start date must be within 3 months of the inclusion notice submission date');
-
-      aboutFacilityUnissued.issueDateDay().clear().type(dateConstants.threeDaysDay);
-      aboutFacilityUnissued.issueDateMonth().clear().type(dateConstants.threeDaysMonth);
-      aboutFacilityUnissued.issueDateYear().clear().type(dateConstants.threeDaysYear);
-
-      aboutFacilityUnissued.shouldCoverStartOnSubmissionNo().click();
-      aboutFacilityUnissued.coverStartDateDay().clear().type(dateConstants.twoMonthsDay);
-      aboutFacilityUnissued.coverStartDateMonth().clear().type(dateConstants.twoMonthsMonth);
-      aboutFacilityUnissued.coverStartDateYear().clear().type(dateConstants.twoMonthsYear);
-
-      aboutFacilityUnissued.coverEndDateDay().clear().type(dateConstants.threeMonthsOneDayDay);
-      aboutFacilityUnissued.coverEndDateMonth().clear().type(dateConstants.threeMonthsOneDayMonth);
-      aboutFacilityUnissued.coverEndDateYear().clear().type(dateConstants.threeMonthsOneDayYear);
 
       if (facilityEndDateEnabled) {
         aboutFacilityUnissued.isUsingFacilityEndDateYes().click();
@@ -276,9 +248,9 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
     });
 
     it('change links should appear for facility four and three should be unissued still', () => {
-      const issuedDate = format(dateConstants.threeDaysAgo, 'd MMMM yyyy');
-      const coverStart = format(dateConstants.twoMonths, 'd MMMM yyyy');
-      const coverEnd = format(dateConstants.threeMonthsOneDay, 'd MMMM yyyy');
+      const issuedDate = format(dateConstants.today, 'd MMMM yyyy');
+      const coverStartNow = format(dateConstants.today, 'd MMMM yyyy');
+      const coverEnd = format(dateConstants.threeYears, 'd MMMM yyyy');
 
       applicationPreview.reviewFacilityStage().contains('Review facility stage');
       applicationPreview.updatedUnissuedFacilitiesHeader().contains('The following facility stages have been updated to issued:');
@@ -294,14 +266,14 @@ context('Unissued Facilities MIN - change to issued from preview page', () => {
       applicationPreview.facilitySummaryListTable(0).nameAction().contains('Change');
       applicationPreview.facilitySummaryListTable(0).issueDateValue().contains(issuedDate);
       applicationPreview.facilitySummaryListTable(0).issueDateAction().contains('Change');
-      applicationPreview.facilitySummaryListTable(0).coverStartDateValue().contains(coverStart);
+      applicationPreview.facilitySummaryListTable(0).coverStartDateValue().contains(coverStartNow);
       applicationPreview.facilitySummaryListTable(0).coverStartDateAction().contains('Change');
       applicationPreview.facilitySummaryListTable(0).coverEndDateValue().contains(coverEnd);
       applicationPreview.facilitySummaryListTable(0).coverEndDateAction().contains('Change');
 
       if (facilityEndDateEnabled) {
-        applicationPreview.facilitySummaryListTable(0).isUsingFacilityEndDateAction().contains('Change');
         applicationPreview.facilitySummaryListTable(0).isUsingFacilityEndDateValue().contains('Yes');
+        applicationPreview.facilitySummaryListTable(0).isUsingFacilityEndDateAction().contains('Change');
       }
 
       // facility three still unissued
